Added new playground wrapper
