+++ apps/web/src/app/[locale]/(playgrounds)/challenge-playground/layout.tsx
@@ -2,9 +2,7 @@ import { ForceRenderUntilClient } from '@repo/ui/components/force-render-until-c
 import { getServerAuthSession } from '@repo/auth/server';
 import { isAdminOrModerator } from '~/utils/auth-guards';
 import { redirect } from 'next/navigation';
-import { useChallengePlaygroundStore } from './challenge-playground-store';
-import { ChallengeLayoutWrapper } from '../../challenge/_components/challenge-layout-wrapper';
-import type { ChallengeRouteData } from '../../challenge/[id]/getChallengeRouteData';
+import { ChallengePlaygroundWrapper } from './ChallengePlaygroundWrapper';
 
 export default async function LayoutData({ children }: { children: React.ReactNode }) {
   const session = await getServerAuthSession();
@@ -12,13 +10,9 @@ export default async function LayoutData({ children }: { children: React.ReactNo
     return redirect('/');
   }
 
-  const { values } = useChallengePlaygroundStore();
-
-  const challenge = values.challenge as ChallengeRouteData;
-
   return (
     <ForceRenderUntilClient>
-      <ChallengeLayoutWrapper challenge={challenge}>{children}</ChallengeLayoutWrapper>
+      <ChallengePlaygroundWrapper>{children}</ChallengePlaygroundWrapper>
     </ForceRenderUntilClient>
   );
 }
