--- apps/web/src/app/[locale]/(playgrounds)/challenge-playground/left-wrapper.tsx
@@ -1,11 +1,0 @@
-'use client';
-
-import { type ReactNode } from 'react';
-
-interface Props {
-  children: ReactNode;
-}
-
-export function LeftWrapper({ children }: Props) {
-  return <div className="flex h-full w-full flex-col">{children}</div>;
-}
